requirements.txt
pillow==12.2.0
rembg==2.0.74
onnxruntime==1.24.4
numpy==2.4.4
pyinstaller==6.19.0
